Add files via upload
This project marks the first iteration of Pawtion Care, developed as a web-based platform focused on providing pet care resources and services. It served as the foundational version that later inspired the full-featured mobile app. The website offers essential features for pet owners to access care information and manage pet needs efficiently online.